arquivos novos

+++ src/main/java/br/com/ifpe/oxefood/api/produto/ProdutoRequest.java
@@ -19,7 +19,7 @@ public class ProdutoRequest {
  
     private String descricao;
  
-    private int valorUnitario;
+    private double valorUnitario;
  
     private int tempoEntrega;
